% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{launchWorkPackageExplorer}
\alias{launchWorkPackageExplorer}
\title{Launch the WorkPackages Shiny app}
\usage{
launchWorkPackageExplorer(
  connectionDetails = NULL,
  databaseSchema = NULL,
  dataFile = NULL,
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package, specifying how to connect to the server where
the study results results have been uploaded using the
\code{\link{uploadResults}} function.}

\item{databaseSchema}{The schema on the database server where the results
have been uploaded.}

\item{dataFile}{(Optional) The name of the .RData file with results. It is commonly known as the
Premerged file.}

\item{runOverNetwork}{(optional) Do you want the app to run over your network?}

\item{port}{(optional) Only used if \code{runOverNetwork} = TRUE.}

\item{launch.browser}{Should the app be launched in your default browser, or in a Shiny window.
Note: copying to clipboard will not work in a Shiny window.}
}
\description{
Launch the WorkPackages Shiny app
}
\details{
Launches a Shiny app that allows the user to explore WorkPackages
}
\examples{
connectionDetails <- DatabaseConnector::createConnectionDetails(
  dbms = "postgresql",
  server = paste(keyring::key_get("workPackagesServer"),
              keyring::key_get("workPackagesDatabase"),
              sep = "/"),
  user = keyring::key_get("workPackagesReadOnlyUser"),
  password = keyring::key_get("workPackagesReadOnlyPassword"))

launchWorkPackageExplorer(connectionDetails = connectionDetails,
                          databaseSchema = paste0(keyring::key_get("workPackagesDatabase"), ".work_packages"))

}
