% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkPackages.R
\docType{package}
\name{WorkPackages-package}
\alias{WorkPackages}
\alias{WorkPackages-package}
\title{WorkPackages: Work packages for on-boarding OHDSI community members}
\description{
A collection of work packages and associated tools for
  on-boarding OHDSI community members
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/WorkPackages/}
}

}
\keyword{internal}
